+++ src/test/java/io/github/sequelcore/vigil/autoconfigure/VigilAutoConfigurationTest.java
@@ -137,6 +137,7 @@ class VigilAutoConfigurationTest {
         null,
         null,
         null,
+        null,
         null);
   }